New Player UI is In
I created a new UI for howler and have started to move almost
everything that is purely jukebox related to the jukebox.css and ~.js
files. My idea is that eventually if anyone were to take this and adapt
it. They could take my jukebox html/css/js and slap it onto a page and
have it work if you feed it a song.