New Functionalities, Stable Release
If there ever were a first release for this, this would be it.
We have a lot of the core functionalities - playing music (of course),
adding songs to the queue and loading in new songs from across a
client’s filesystem. The album artwork is loaded in and if it’s not
found I use siriwave to showoff something instead of nothing. The
entire library can be added to the playlist by the addAll functions.
Through different sessions, the manifest.json file persists and will
relocate the music upon reopening the application letting users be able
to add once, use forever (until they move their music around in their
filesystem at least!). By all these functions combined, I give you the
first release.