General Tuneups/Networking
Certain screws were just loose in the system and I’ve spent a little
bit of time tightening down the hatches.

New with this is the removal of the Channel Users tab. This is reserved
for when a channel is actually hosted. Otherwise, there should only be
one user - you and that’s not interesting enough to show.

Otherwise, some of the networking stuff has begun to be worked on. But
it’s still a long way from being complete.