More Efficiency and Metadata
Continuing to build on the efficiency in the last patch, I just made
the library have the ability to append to itself again. Much much
faster than before since the collection is now id based and the DOM
elements contain their lookup ids.

Metadata is on the horizon, and with the notify.js library we can
cleanly use notification windows to prompt the user for data switching.

Next will be to make the whole metadata changing system more robust and
bug free because there’s a few pesky bugs that I could not initially
get out of there.

Also ‘var’s are now ‘let’s.