Redoing Append to Library + Testing
Append to library used to run checks by iterating over the song’s name.
Now we do everything by ID. Ytdl is working again too!

Also most of the testing for the tree.js is in. Still working on
wrapping up the last few functions but it’s mostly there.